Refactored DataGenerator (into SegmentationGenerator) and ClassificationGenerator so that they work now
Changed up variable names for consistency purposes and remoevd unnecessary resizing. Also, I had to resort to the np.unique way of creating classification labels because in the current pipeline, some of the encoded pixels are pandas Sequences and parsing them into proper rle labels is a lot of work (for future endeavors).