Fixed bug with stacking the ensembled arrays and added tqdm
The ensembling stacking bug was how np.mean was averaging across the number of x-rays instead of the number of ensembled images. This caused the shape to go from (n, h, w) to just (h,w) when we want (n, h, w). The fix was to change from np.vstack to np.stack so that the images go from (n_ensemble, n, h, w) to (n, h, w). Also I added tqdm for better UI/UX.