Can now save the classification probabilities as a .csv file and also made the thresholding + df creation a separate function and fixed TTA for classification
The probabilities can be saved using the `save_p` option. Also made a new function `create_thresholded_classification_csv()` to save the thresholded classification predictions as a .csv file for increased modularity. It's also for parallelism purposes because it looks cleaner with it and the `create_classification_p_df()` function.
The bugs with TTA_Classification came from misnamed arguments (n_iter) and how data_augmentation returns a tuple of arrs instead of just a single arr.