Created a pipeline for segmentation-only inference
Similar to Stage2 in segmentation.py but is heavily refactored and much cleaner. The code is a bit more modular (run_seg_prediction, create_sub_from_rles) and since it is using the whole test dataset, I had to implement a way to be memory-conscious. To do so, I added the argument, fpaths_batch_size, which lets you specify how to batch the incoming test_fpaths so that you don't have to load in the entire dataset to predict on it. This comes in handy when ensembling and using TTA, where there's going to multiple arrays the size of the entire test set in the memory at once. Also the main loop is much more different, where rles are straight up calculated instead of in a separate wrapper around mask2rle.