Added the arg tta_then_preprocess to allow the user to specify the order of TTA/preprocessing
This was to address the issue where grayscale models were being trained with [TTA->preprocess_input] instead of the inference pipeline's [preprocess_input->TTA]. It's also cleaner this way because np.invert is supposed to only work with integers in [0, 255], not the preprocesed/resized inputs. The current implementation is pretty hacky, so it'll need some refactoring in the future, but it's fine for now.