Refactored functions so that there is less repeat code with preprocess_fn in the arguments
Reduced the number of **kwargs occurrences in the functions arguments (only Stage1 has it) by making it so that Stage1 has it made into a partial'd function. Fixed documentation to accomodate and added support in TTA_Classification to handle cases where preprocess_fn is None.
Also, fixed a bug where batch_test_fpaths was called instead of test_fpaths_batched.